refactor: migrate middleware functions to services package for better organization
