+++ ee/sso/service.go
@@ -14,7 +14,6 @@ import (
 	"expo-open-ota/ee/licensing"
 	"expo-open-ota/internal/auditlog"
 	"expo-open-ota/internal/crypto"
-	"expo-open-ota/internal/middleware"
 	"expo-open-ota/internal/services"
 	"expo-open-ota/internal/store"
 	"fmt"
@@ -509,7 +508,7 @@ func (s *SSOService) recordConfigEvent(ctx context.Context, action auditlog.Acti
 		return
 	}
 	actorID, actorDisplay := "", ""
-	if principal := middleware.PrincipalFromContext(ctx); principal != nil {
+	if principal := services.PrincipalFromContext(ctx); principal != nil {
 		actorID, actorDisplay = principal.UserId, principal.Email
 		if actorDisplay == "" {
 			actorDisplay = principal.UserId
